Item Categorization. (#68)

+++ src/SkillCraft.Api.Core/Items/Weight.cs
@@ -5,15 +5,15 @@ namespace SkillCraft.Api.Core.Items;
 
 public class Weight
 {
-  public double Value { get; }
+  public int Value { get; }
 
-  public Weight(double value)
+  public Weight(int value)
   {
     Value = value;
     new Validator().ValidateAndThrow(this);
   }
 
-  public static Weight? TryCreate(double? value) => value.HasValue ? new(value.Value) : null;
+  public static Weight? TryCreate(int? value) => value.HasValue ? new(value.Value) : null;
 
   public override bool Equals(object? obj) => obj is Weight weight && weight.Value.Equals(Value);
   public override int GetHashCode() => Value.GetHashCode();
