+++ src/SkillCraft.Api.Infrastructure/Converters/WeightConverter.cs
@@ -6,7 +6,7 @@ internal class WeightConverter : JsonConverter<Weight>
 {
   public override Weight? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetDouble(out double value) ? new Weight(value) : null;
+    return reader.TryGetInt32(out int value) ? new Weight(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, Weight weight, JsonSerializerOptions options)
